// fz-backend/src/analysis/interfaces/deepseek-response.interface.ts
// src/modules/analysis/interfaces/deepseek-response.interface.ts

export type DeepSeekRole = 'system' | 'user' | 'assistant';

export interface DeepSeekMessage {
    role: DeepSeekRole; // ← используем конкретный тип, а не string
    content: string;
}

export interface DeepSeekRequest {
    model: string;
    messages: DeepSeekMessage[];
    temperature?: number;
    max_tokens?: number;
    reasoning_effort?: 'low' | 'medium' | 'high';
    thinking?: { "type": "enabled" },
    stream?: false,
}

export interface DeepSeekResponse {
    id: string;
    choices: Array<{
        index: number;
        message: DeepSeekMessage;
        finish_reason: string;
    }>;
    usage: {
        prompt_tokens: number;
        completion_tokens: number;
        total_tokens: number;
    };
}

export interface UploadedFile {
    fileId: string;
    fileName: string;
    bytes: number;
}